House platforms done
Not really sure if there's any platform work I can do inside so this level's done platform-wise (I still need to add details)
There's an edge collider called Top, which links to the top platform that you cannot go back from once you jump through it. This is intentional, as it's meant to be a boss room. If we don't need it, I can delete it later.