I've got an idea
It's a crazy idea, so I'm going to finish level 5 (which won't use it) before I work on said crazy idea in level 4.
Basically, every level I'm making uses a gimmick, except level 1, which is the tutorial. 2 has some interesting spike sections, 3 introduces portals and has multiple paths, and 5 is going to be dark, with bare outlines highlighting the platforms. Level 4, my crazy idea, is going to have random portals that shunt the player around the level randomly, and then eventually into the boss room. Basically it's going to be a series of gauntlets, and I'll add a timer that sets when you can go to the boss level (after a certain amount of portals). What do you guys think?
I know you don't read these descriptions but just in case lol